warns PVS

+++ sources/DLoader/src/FlashDrive/FlashDrive.cpp
@@ -411,7 +411,7 @@ void Upgrade()
         size -= readedBytes;
         address += static_cast<uint>(readedBytes);
 
-        percentsUpdate = 1.0F - static_cast<float>(size) / fullSize;
+        percentsUpdate = 1.0F - static_cast<float>(size) / static_cast<float>(fullSize);
     }
 
     FDrive::CloseOpenedFile();
